return dc

+++ ROMS/Model/TRN_ReturnDC.cs
@@ -31,14 +31,14 @@ namespace ROMS.Model
         public int ParaSubGroupID { get; set; } = 0;
         public int paraReasonId { get; set; } = 0;
         public int paraClosingReasonId { get; set; } = 0;
-        public double paraReturnDCAmount { get; set; } = 0;
+        public decimal paraReturnDCAmount { get; set; } = 0;
         public string paraCreditNoteNo { get; set; } = "";
         public string paraExchangeRemarks { get; set; } = "";
         public string paraCreditNoteDate { get; set; } = "";
         public string paraRemarks { get; set; } = "";
         public double paraDeleteFlag { get; set; } = 0;
-        public  double ParaSubtotal { get; set; } = 0;
-        public double paraTax { get; set; } = 0;
+        public  decimal ParaSubtotal { get; set; } = 0;
+        public decimal paraTax { get; set; } = 0;
         public DataTable paraTRN_Purchase_ReturnDC { get; set; } = null;
         public DataTable @ParaTRN_Purchase_DC { get; set; } = null;
     }
